refactor parser to directory module
