+++ cobalt-parser/src/parser/expr/mod.rs
@@ -1,0 +1,61 @@
+use std::borrow::Cow;
+
+use cobalt_ast::{ast::VarGetAST, BoxedAST};
+use cobalt_errors::{CobaltError, ParserFound, SourceSpan};
+
+use crate::lexer::tokens::TokenKind;
+
+use super::Parser;
+
+mod literal;
+
+impl<'src> Parser<'src> {
+    pub fn parse_expr(&mut self) -> (BoxedAST<'src>, Vec<CobaltError<'src>>) {
+        assert!(self.current_token.is_some());
+
+        match self.current_token.unwrap().kind {
+            TokenKind::Literal(_) => {
+                return self.parse_literal();
+            }
+            TokenKind::Ident(_) => {
+                return self.parse_ident_expr();
+            }
+            _ => {}
+        }
+
+        (
+            Box::new(cobalt_ast::ast::NullAST::new(SourceSpan::from((0, 1)))),
+            vec![CobaltError::ExpectedFound {
+                ex: "expression",
+                found: ParserFound::Str(self.current_token.unwrap().kind.to_string()),
+                loc: self.current_token.unwrap().span,
+            }],
+        )
+    }
+
+    fn parse_ident_expr(&mut self) -> (BoxedAST<'src>, Vec<CobaltError<'src>>) {
+        assert!(self.current_token.is_some());
+
+        let errors = vec![];
+
+        let span = self.current_token.unwrap().span;
+        let name = match self.current_token.unwrap().kind {
+            TokenKind::Ident(name) => Cow::from(name),
+            _ => {
+                return (
+                    Box::new(cobalt_ast::ast::NullAST::new(SourceSpan::from((0, 1)))),
+                    vec![CobaltError::ExpectedFound {
+                        ex: "identifier",
+                        found: ParserFound::Str(self.current_token.unwrap().kind.to_string()),
+                        loc: span,
+                    }],
+                );
+            }
+        };
+        let is_global = false;
+
+        self.next();
+
+        return (Box::new(VarGetAST::new(span, name, is_global)), errors);
+    }
+}
